backend/requirements.txt
django
djongo
djangorestframework
django-cors-headers
pyjwt